[CP-2425] Extracted history properties (HistoryLocationState) in start
