+++ libs/core/core/errors/app-error.ts
@@ -6,7 +6,7 @@
 export class AppError<Type extends string = string> extends Error {
   constructor(
     public readonly type: Type,
-    public readonly message: string,
+    public readonly message: string = "",
     // AUTO DISABLED - fix me if you like :)
     // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types, @typescript-eslint/no-explicit-any
     public readonly payload?: any
