+++ libs/core/device-initialization/actions/should-skip-data-sync.helper.ts
