+++ libs/core/device-initialization/constants/errors.enum.ts
@@ -3,7 +3,6 @@
  * For licensing, see https://github.com/mudita/mudita-center/blob/master/LICENSE.md
  */
 
-export enum PasscodeLockedTestIds {
-  Timer = "passcode-timer",
-  Container = "phone-locked-container",
+export enum DeviceInitializationError {
+  InitializingDeviceError = "device-initialization-initializing-device-error",
 }
