+++ libs/core/device-initialization/reducers/device-initialization.reducer.ts
@@ -9,7 +9,6 @@ import {
   DeviceInitializationStatus,
 } from "Core/device-initialization/reducers/device-initialization.interface"
 import { setDeviceInitializationStatus } from "Core/device-initialization/actions/base.action"
-import { startInitializingDevice } from "Core/device-initialization/actions/start-initializing-device/start-initializing-device.action"
 
 export const initialState: DeviceInitializationState = {
   deviceInitializationStatus: DeviceInitializationStatus.Idle,
@@ -17,17 +16,10 @@ export const initialState: DeviceInitializationState = {
 
 export const deviceInitializationReducer =
   createReducer<DeviceInitializationState>(initialState, (builder) => {
-    builder
-      .addCase(setDeviceInitializationStatus, (state, action) => {
-        return {
-          ...state,
-          deviceInitializationStatus: action.payload,
-        }
-      })
-      .addCase(startInitializingDevice.pending, (state) => {
-        return {
-          ...state,
-          deviceInitializationStatus: DeviceInitializationStatus.Initializing,
-        }
-      })
+    builder.addCase(setDeviceInitializationStatus, (state, action) => {
+      return {
+        ...state,
+        deviceInitializationStatus: action.payload,
+      }
+    })
   })
